+++ packages/in-memory/tests/inMemory.spec.ts
@@ -5,7 +5,7 @@ import {
   navigateSuite,
   goSuite,
   cancelSuite,
-  blockingSuite
+  confirmationSuite
 } from "../../../tests/cases";
 
 import { TestCase, Suite } from "../../../tests/types";
@@ -142,8 +142,8 @@ describe("go suite", () => {
   runSuite(goSuite);
 });
 
-describe("blocking suite", () => {
-  runSuite(blockingSuite);
+describe("confirmation suite", () => {
+  runSuite(confirmationSuite);
 });
 
 describe("go", () => {