+++ packages/root/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Next
 
-- Add navigation blocking to each history type. Blocker is added using `history.confirm(fn)` and removed using `history.confirm()` (no argument).
+- Add navigation confirmation to each history type. Confirmation function is added using `history.confirm(fn)` and removed using `history.confirm()` (no argument).
 - Rename `navigationConfirmation` to `confirmation`. `navigationConfirmation` still exists, but will be removed in next major.
 
 ## 2.0.0